+++ src/lp-solve/read.h
@@ -9,8 +9,8 @@ void store_re_op(void);
 void null_tmp_store(void);
 void store_bounds(void);
 void add_int_var(char *name);
-void rhs_store(REAL value);
-void var_store(char *var, int row, REAL value);
+void rhs_store(double value);
+void var_store(char *var, int row, double value);
 
 /* This one seems to have been forgotten in the original code.  Added
    4/15/2003 by arc */
